fix(network): filter out localhost when identify notifies kademlia
