+++ crates/papyrus_network/src/network_manager/mod.rs
@@ -1,5 +1,4 @@
 mod swarm_trait;
-use core::net::Ipv4Addr;
 
 #[cfg(test)]
 mod test;
@@ -14,10 +13,9 @@ use futures::future::{ready, BoxFuture, Ready};
 use futures::sink::With;
 use futures::stream::{self, FuturesUnordered, Map, Stream};
 use futures::{pin_mut, FutureExt, Sink, SinkExt, StreamExt};
-use libp2p::core::multiaddr::Protocol;
 use libp2p::gossipsub::{SubscriptionError, TopicHash};
 use libp2p::swarm::SwarmEvent;
-use libp2p::{Multiaddr, PeerId, StreamProtocol, Swarm};
+use libp2p::{PeerId, StreamProtocol, Swarm};
 use metrics::gauge;
 use papyrus_common::metrics as papyrus_metrics;
 use sqmr::Bytes;
@@ -28,7 +26,7 @@ use crate::bin_utils::build_swarm;
 use crate::gossipsub_impl::Topic;
 use crate::mixed_behaviour::{self, BridgedBehaviour};
 use crate::sqmr::{self, InboundSessionId, OutboundSessionId, SessionId};
-use crate::utils::StreamHashMap;
+use crate::utils::{is_localhost, StreamHashMap};
 use crate::{gossipsub_impl, NetworkConfig};
 
 #[derive(thiserror::Error, Debug)]
@@ -923,15 +921,4 @@ pub struct BroadcastNetworkMock<T: TryFrom<Bytes>> {
 pub struct TestSubscriberChannels<T: TryFrom<Bytes>> {
     pub subscriber_channels: BroadcastSubscriberChannels<T>,
     pub mock_network: BroadcastNetworkMock<T>,
-}
-
-fn is_localhost(address: &Multiaddr) -> bool {
-    let maybe_ip4_address = address.iter().find_map(|protocol| match protocol {
-        Protocol::Ip4(ip4_address) => Some(ip4_address),
-        _ => None,
-    });
-    let Some(ip4_address) = maybe_ip4_address else {
-        return false;
-    };
-    ip4_address == Ipv4Addr::LOCALHOST
 }
